Extracted the classes from the ExampleMod.java Moved the tool templates to own package tools.templates and the tools to package tools.
