+++ src/main/java/net/fabricmc/example/tools/RubyWeapon.java
@@ -1,5 +1,6 @@
-package net.fabricmc.example;
+package net.fabricmc.example.tools;
 
+import net.fabricmc.example.ExampleMod;
 import net.minecraft.item.ToolMaterial;
 import net.minecraft.recipe.Ingredient;
