+++ src/main/java/net/fabricmc/example/tools/templates/CustomAxeItem.java
@@ -1,0 +1,10 @@
+package net.fabricmc.example.tools.templates;
+
+import net.minecraft.item.AxeItem;
+import net.minecraft.item.ToolMaterial;
+
+public class CustomAxeItem extends AxeItem {
+    public CustomAxeItem(ToolMaterial material, int attackDamage, float attackSpeed, Settings settings) {
+        super(material, attackDamage, attackSpeed, settings);
+    }
+}
