+++ src/main/java/net/fabricmc/example/tools/templates/CustomHoeItem.java
@@ -1,0 +1,10 @@
+package net.fabricmc.example.tools.templates;
+
+import net.minecraft.item.HoeItem;
+import net.minecraft.item.ToolMaterial;
+
+public class CustomHoeItem extends HoeItem {
+    public CustomHoeItem(ToolMaterial material, int attackDamage, float attackSpeed, Settings settings) {
+        super(material, attackDamage, attackSpeed, settings);
+    }
+}
